feat(SDK): add basic support for Oculus SDK and OVRCameraRig
  > ** Breaking Changes **

The native Oculus SDK provides a prefab called OVRCameraRig and this
feature now provides basic support of this prefab within VRTK.

The Haptics have been updated to no longer reference the SteamVR
strength of 0 to 3999, but now use a float between 0f and 1f to
determine the strength of the rumble. The Oculus SDK also uses 0f to 1f
so this is a simple pass through, but the SteamVR SDK converts the 0f
or 1f strength into a value between 0 and 3999.

+++ Assets/VRTK/Scripts/Controls/2D/RadialMenu/VRTK_IndependentRadialMenuController.cs
@@ -222,7 +222,7 @@ namespace VRTK
             }
         }
 
-        protected override void AttemptHapticPulse(ushort strength)
+        protected override void AttemptHapticPulse(float strength)
         {
             if (interactingObjects.Count > 0)
             {
